# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: landmark area API
  description: landmark service
  version: "1.0"
paths:
...
  /administrative-areas/{id}/simple:
    get:
      summary: display single administrative area (without geometry)
      operationId: loadAdministrativeAreaSimple
      tags:
        - administrativeAreaSimple
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministrativeAreaResponse"
  /administrative-areas/by-level-country/simple:
    get:
      summary: Get administrative areas by level and country name (simple, without geometry)
      operationId: getAdministrativeAreasByLevelAndCountrySimple
      tags:
        - administrativeAreaSimple
      parameters:
        - in: query
          name: levelName
          required: true
          schema:
            type: string
        - in: query
          name: countryName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of administrative areas (without geometry)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdministrativeAreaResponse"
  /administrative-areas/by-metadata/simple:
    get:
      summary: Get administrative areas by metadata query (simple, without geometry)
      operationId: getAdministrativeAreasByMetadataSimple
      tags:
        - administrativeAreaSimple
      parameters:
        - in: query
          name: metadataQuery
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of administrative areas (without geometry)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdministrativeAreaResponse"
  /administrative-levels:
    post:
      summary: Create level
      operationId: createAdministrativeLevel
      tags:
        - administrativeLevel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministrativeLevel"
        required: true
      responses:
        "400":
          description: Bad Request
        "201":
          description: Created
  /administrative-levels/{AdministrativeLevelId}:
    get:
      summary: display single administrative level
      operationId: loadAdministrativeLevel
      tags:
        - administrativeLevel
      parameters:
        - in: path
          name: AdministrativeLevelId
          required: true
          schema:
            type: AdministrativeLevelId
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministrativeLevel"
    put:
      summary: update single administrative level
      operationId: updateAdministrativeLevel
      tags:
        - administrativeLevel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministrativeLevel"
        required: true
      parameters:
        - in: path
          name: AdministrativeLevelId
          required: true
          schema:
            type: AdministrativeLevelId
      responses:
        "400":
          description: Bad Request
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdministrativeLevel"
    delete:
      summary: delete single administrative level
      operationId: deleteAdministrativeLevel
      tags:
        - administrativeLevel
      parameters:
        - in: path
          name: AdministrativeLevelId
          required: true
          schema:
            type: AdministrativeLevelId
      responses:
        "400":
          description: Bad Request
        "204":
          description: OK
...
components:
  schemas:
...
    AdministrativeAreaResponse:
      type: object
      required:
        - name
      properties:
        id:
          type: AdministrativeAreaId
        name:
          type: LandmarkName
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        level:
          $ref: "#/components/schemas/AdministrativeLevel"
        parent:
          type: AdministrativeAreaId
        geometry:
          type: string
        metadata:
          $ref: "#/components/schemas/LandmarkMetadata"
...
    AdministrativeLevel:
      type: object
      required:
        - order
        - country
        - name
      properties:
        id:
          type: AdministrativeLevelId
        name:
          type: LevelName
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        country:
          type: CountryName
        order:
          type: LevelOrder
...
    LandmarkMetadata:
      type: object